restructured components folder
