+++ client/src/app/Search/layout.tsx
@@ -1,0 +1,13 @@
+import { Box } from "@chakra-ui/react";
+
+export default function SearchPageLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <Box>
+      <Box>{children}</Box>
+    </Box>
+  );
+}
