+++ client/src/app/Search/page.tsx
@@ -1,0 +1,13 @@
+"use client";
+import { Heading } from "@chakra-ui/react";
+import React from "react";
+
+export default function SearchPage() {
+  return (
+    <>
+      <Heading as="h1" size="2xl" mb={4}>
+        Search Page
+      </Heading>
+    </>
+  );
+}
