+++ client/src/components/HomePage/HelpCard.tsx
