+++ client/src/components/HomePage/HelpSection.tsx
