+++ client/src/components/HomePage/HeroSVG.tsx
