+++ client/src/components/HomePage/HeroSection.tsx
